SmartCoCo/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
black==23.3.0
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
crytic-compile==0.3.1
cytoolz==0.12.1
dbus-python==1.2.18
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
func-timeout==4.3.5
hexbytes==0.3.0
idna==3.4
joblib==1.2.0
jsonschema==4.17.3
Levenshtein==0.21.0
lru-dict==1.1.8
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parsimonious==0.9.0
pathspec==0.11.1
platformdirs==3.5.0
prettytable==3.7.0
protobuf==4.22.3
py-solc-x==1.1.1
pycryptodome==3.17
PyGObject==3.42.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
slither-analyzer==0.9.3
solc-select==1.0.3
solidity-parser==0.1.1
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
version-parser==1.0.1
wcwidth==0.2.6
web3==6.3.0
websockets==11.0.2
yarl==1.9.2
